requirements.txt
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
exceptiongroup==1.3.0
flet==0.28.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==6.0.0
oauthlib==3.3.1
pdf2image==1.17.0
Pillow==9.0.1
pypandoc==1.15
pypdf==5.8.0
python-docx==1.2.0
repath==0.9.0
reportlab==4.4.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.14.1
